Source the windows service name from setenv
You can use the variable SERVICE_NAME to change the service name. But keep in mind to no use whitespaces in the name.

Added a backslash to the BLENDED_HOME variable to work around some quoting issues in it's later use.

Changed path to jvm.dll from server to client directory.